refactor(tests): split CI execution-policy test monolith by owner (#2002)
Split test_ci_execution_policy.py (470 lines, 32 tests) into owner-local
files: test_product_ci_contract.py (13 tests), test_benchmark_workflow_contract.py
(11 tests), test_make_commands.py (4 tests), and test_precommit_entrypoints.py
(1 test). A benchmark-workflow policy change and a product-wheel/Lean change
now produce disjoint test diffs.
